+++ src/main/java/com/dmf/loja/validation/documento/CPFCNPJ.java
@@ -8,7 +8,7 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-@Constraint(validatedBy = DocumentoValidator.class)
+@Constraint(validatedBy = DocumentoValidatorProprio.class)
 @Target({ ElementType.FIELD })
 @Retention(RetentionPolicy.RUNTIME)
 public @interface CPFCNPJ {
